Fill readme, create base class predictor
