+++ README.md
@@ -7,3 +7,15 @@
 Описание модуля `translator` (translator/README.md), В нем же приведено описание моделей и методов оценки модели.
 
 
+В самом простом варианте реализация переводчика занимает несколько строк:
+
+```python
+from translator.predictor import CommonPredictor
+from translator.translator import OPUSModel, DEFAULT_PRETRAINED_USER_MODEL_FILEPATH
+
+predictor = CommonPredictor(OPUSModel, DEFAULT_PRETRAINED_USER_MODEL_FILEPATH)
+
+predicted_text = predictor.predict("My name is Nicklai and I live in Almetievsk.")
+
+print(f"Translation: {predicted_text}")
+```
