+++ translator/README.md
@@ -1,0 +1,9 @@
+# Translator
+
+Модуль `translator`
+
+Написали класс `BaseModel` для моделей переводчиков, чтобы можно было Приведенную модель можно расширять на  с возможностью использования 
+
+
+
+В репозитории [`huggingface/transformers`](https://github.com/huggingface/transformers) есть описание приве
